requirements.txt
# Core FastAPI
fastapi==0.121.2
uvicorn[standard]==0.38.0
greenlet==3.2.4
pydantic==2.12.4
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.17.2

# Redis
redis==5.2.1

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.20

# Data validation and serialization
email-validator==2.3.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
PyJWT==2.10.1

# Development
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1

# Code quality
black==25.11.0
flake8==7.3.0
isort==7.0.0
mypy==1.18.2
pylint==4.0.3

# Monitoring/Logging
python-logging-loki==0.3.1
